format-patch: don't leak "extra_headers" or "ref_message_ids"
Fix two memory leaks in "struct rev_info" by freeing that memory in
cmd_format_patch(). These two are unusual special-cases in being in
the "struct rev_info", but not being "owned" by the code in
revision.c. I.e. they're members of the struct so that this code in
"builtin/log.c" can pass information code in log-tree.c.

See 20ff06805c6 (format-patch: resurrect extra headers from config,
2006-06-02) and d1566f7883f (git-format-patch: Make the second and
subsequent mails replies to the first, 2006-07-14) for the initial
introduction of "extra_headers" and "ref_message_ids".

We can count on repo_init_revisions() memset()-ing this data to 0
however, so we can count on it being either NULL or something we
allocated. In the case of "extra_headers" let's add a local "char *"
variable to hold it, to avoid the eventual cast from "const char *"
when we free() it.

While we're at it let's also move to using string_list_init_nodup()
instead of relying on calloc() implicitly coming up with the same
result. See 770fedaf9fb (string-list.[ch]: add a
string_list_init_{nodup,dup}(), 2021-07-01) for the introduction of
the string_list_init_nodup() helper. The string_list_clear() here is
redundant due to the *_nodup() initialization, but let's add it anyway
for consistency with other API use.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>