revisions API users: use release_revisions() in wt-status.c edge case
Use release_revisions() for the "struct rev_info" in
has_unstaged_changes() which didn't require the addition of a "goto"
pattern, but needed us to refactor another "return fn(...)" pattern.

This change is split from preceding changes to make similar
conversions in those commits easier to review, as they are all
repetitions of the same pattern.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>