tests: stop using test_cmp_config "" --default "" <key> pattern
Using the '"" --default ""' pattern introduced in d3115660b4c (branch:
add flags and config to inherit tracking, 2021-12-20) made can be
replaced by the more straightforward and accurate testing of the "git
config" return value for that key.

All of these tests were trying to assert that the configuration didn't
exist, but instead they'd conflate nonexistence with the key being set
to the empty string.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>