ls-tree: remove FIELD_*, just use MODE_*
When we're picking where we should go in the optimized "show_tree"
path there's no reason for why we need to convert our "cmdmode" of
e.g. MODE_LONG into a FIELD_LONG_DEFAULT. Instead we can simply do
those checks in the show_tree() function itself.

Let's also make this code more future-proof by unrolling the hardcoded
strmp() if/else if chain into something that checks a new "static
struct" providing a bidirectional mapping between optimized formats
and the ls_tree_cmdmode.

A preceding commit removed "MODE_UNSPECIFIED", now there is a use for
such a thing, as we'll need to add this to our mapping so we'll choose
the "show_tree" fast-path here.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>