i18n: handle "unsigned long" overflow in Q_()'s "n"
Change the Q_() function added in 0c9ea33b90f (i18n: add stub Q_()
wrapper for ngettext, 2011-03-09) to take a "size_t" instead of an
"unsigned long".

For most things in this codebase we'd generally prefer to move to
"size_t" or other types integer types that can represent the maximum
unsigned size on the platform, whereas "unsigned long" is 32 bit
unsigned on some 64 bit systems (notably Windows[1]).

However, in the case of the ngettext() function we don't have that
option. It's a library function whose "n" argument isn't going to
change from "unsigned long".

What we can do is do something about the admittedly obscure case where
that "n" overflows. In this case the "n" is only used by libintl to
determine which plural form it should select, which is done by passing
that "n" to the Plural-Form expression found in the *.po files:

    git grep Plural-Forms -- po

So as unsigned integer overflow is defined in C to wrap-around (unlike
signed overflow, which is undefined), the only edge case we'd get here
is that as we'd e.g. overflow 2^32 we'd start picking the singular
form instead of the plural form. Then as we'd continue to overflow
past say 2^32+100 we'd be pretty much guaranteed to pick the correct
plural form.

This is because most human languages only have plural forms for 0, 1,
>2 etc, or in some rare cases for numbers as high as 20. But we can be
>pretty safe in assuming that almost nobody will want 2^31-1, 2^32 and
>2^31+1 to have different plural forms.

This assumption will not hold for Polish and Russian, whose plural
forms are, respectively:

    "Plural-Forms: nplurals=3; plural=(n==1 ? 0 : n%10>=2 && n%10<=4
    && (n%100<10 || n%100>=20) ? 1 : 2);\n"

    "Plural-Forms: nplurals=4; plural=(n%10==1 && n%100!=11 ? 0 :
    n%10>=2 && n%10<=4 && (n%100<12 || n%100>14) ? 1 : n%10==0
    || (n%10>=5 && n%10<=9) || (n%100>=11 && n%100<=14)? 2 : 3);\n"

I.e. any assumption about being able to check if the number is
divisible by 10 will break down, but that was the case before as we'd
overflow from say 2^32 (4294967296) to 0, then e.g. 2^32+4 would be 4
instead of 4294967300, and not divisible by 100 anymore.

That was the case already, and I don't think we need to worry about
such obscurity.

1. https://en.wikipedia.org/wiki/64-bit_computing#64-bit_data_models

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>