Makefile: create a probe/* tree, use in builtin/bugreport.c
In preparation for using the get_compiler_info() and get_libc_info()
as part of the Makefile move the compat/compiler.h header added in
69bcbbceb74 (bugreport: add compiler info, 2020-04-16) to
probe/{compiler,libc}.[ch].

The interface is intended to use only the C standard library, so that
the Makefile itself can compile a version of these early on to replace
some of the shelling out that it has to do now, e.g. in the
"detect-compiler" script added in 1da1580e4c2 (Makefile: detect
compiler and enable more warnings in DEVELOPER=1, 2018-04-14).

Right now the probe/*.c code includes git-compat-util.h, but in a
subsequent commit macros will be added to it to only use stdlib.h
etc. if we're compiling a stand-alone binary for use by the Makefile
itself.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>