config.mak.dev: support the IBM xlc compiler
Guard the various gcc and clang-specific code in config.mak.dev by
using probe/compiler.c to detect the compiler.

It's now almost possible to compile with xlc on IBM AIX using:

    make DEVELOPER=1 DEVOPTS=no-pedantic

The "almost" part of that is being resolved in the subsequent commit.

We will also (not that I think anyone really cares) report the correct
compiler version in "git bugreport", since it will also use this
probe/compiler.c function as a library function.

Handling "errors" with a mere mere fprintf() here and not returning
non-zero is intentional, we don't want to have the compilation error
out almost immediately because we don't recognize the compiler.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>