parse-options: add per-option flag to stop abbreviation
For cases like "format-patch --output-directory=dir" we disallow
abbreviations entirely, because we'd like setup_revisions() to do a
subsequent option parsing pass, and it accepts "--output=file".

That's one reason for why the parse_options() in diff.c added in
4a288478394 (diff.c: prepare to use parse_options() for parsing,
2019-01-27) passes PARSE_OPT_KEEP_UNKNOWN. I.e. we need the
"--output=file" to be retained in the arguments for passing it down to
setup_revisions().

It's not stated explicitly but I think "--output"
v.s. "--output-directory" is the specific case the author had in mind
when adding the PARSE_OPT_KEEP_UNKNOWN case in the pre-image. That was
added in an earlier commit in the same series,
baa4adc66ae (parse-options: disable option abbreviation with
PARSE_OPT_KEEP_UNKNOWN, 2019-01-27).

As that commit shows we had to disallow "git difftool --symlink" as an
abbreviation for "git difftool --symlinks" as a side-effect, even
though that abbreviation wouldn't have conflicted.

So let's tweak this rule to be more narrowly scoped, now we'll allow
abbreviated options regardless of PARSE_OPT_KEEP_UNKNOWN being set for
parse_options(), but we'll provide individual options an opt-out. By
using that opt-out for "--output-directory" we can have that case work
without overzealously disallowing others.

We can use the "1 << 8" flag here because it was freed up in my
3b723f722df (parse-options.h: move PARSE_OPT_SHELL_EVAL between enums,
2021-09-28).

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>