parse-options: clarify "disallow abbreviated" error messaging
Change the die() added b02e7d5d701 (tests: disallow the use of
abbreviated options (by default), 2019-04-12) to be a BUG()
instead. If we encounter this it's a bug in our test suite, so it
should call BUG() instead of die().

Furthermore when this was enabled the error message we got was less
useful than the one we'd print without it being enabled, since we
wouldn't tell you what option conflicted in the case where
"ambiguous_option" was true.

So let's first print the error(), and then if
GIT_TEST_DISALLOW_ABBREVIATED_OPTIONS is enabled we'll BUG() out, and
have those BUG() messages point to the relevant environment variable.

Since we needed the environment variable as a "const char *" let's
move the static variable into the parse_long_opt() function. There's
no reason we can't defer this, it doesn't need to happen at the start
of parse_options().

As we're returning an exit code test_must_fail would normally not
accept we need to adjust tests I added in fb20d4b1268 (pack-objects
tests: cover blindspots in stdin handling, 2021-06-21). For those
tests just checking exit status 134 and whether we have BUG: on stderr
is enough.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>