object API: rename "object_list" to "object_queue"
Subsequent work will introduce a "struct object_list_NEW" which is
similar to a "struct string_list", but with "struct object *" instead
of a "const char *". The current "struct object_array" does that, but
also tracks "name", "path" and "mode", which as will be demonstrated
the majority of callers don't need.

The problem is that we'll then have "struct
object_{list,array,array_NEW}" in play, and any "array_NEW" I could
come up with that isn't "object_list" is a bad name (the least worst
was "object_array_bare").

So while I don't like the churn I think the least bad option is to
rename "object_list" to "object_queue", which is a better name for it
and its use-cases in any case. As we'll see in subsequent commits some
users of this "object_queue" are really better off with the new
"object_list", i.e. they really just want a bag of objects.

The remaining users of the API left after that will want the semantics
of this API, exactly because this datastructure makes sense for the
users that want a queue-like structure.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>