advice doc: rename from->to advice.{,branch}SetUpstreamFailure
Add a configuration setting for the advice added in
caa2036b3b4 (branch: give advice when tracking start-point is missing,
2013-04-02).

That change did not follow the steps later described in
b3b18d16213 (advice: revamp advise API, 2020-03-02) for adding a new
advice type, but which were already the usual workflow for adding
advice at the time.

While the configuration infrastructure existed and setting
advice.setUpstreamFailure could have suppressed the output, neither
the non-existing documentation nor the message itself referenced that
config key name.

It would thus only have been useful to someone who'd source dived to
discover what the hidden config setting was. Since it was that obscure
it should be safe to rename it while we're at it. By giving it a
"branch" prefix it's immediately obvious what it's for.

This change allows us to remove one of the remaining exceptions in the
"advice_labels_manual" function.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>