advice doc: add missing advice.objectNameWarning section
Add documentation for the advice.objectNameWarning setting added in
798c35fcd8a (get_sha1: warn about full or short object names that look
like refs, 2013-05-29).

That change did not follow the steps later described in
b3b18d16213 (advice: revamp advise API, 2020-03-02) for adding a new
advice type, but which were already the usual workflow for adding
advice at the time.

This allows us to remove the last remaining exceptions in the
"advice_labels_manual" function, which allows us to remove that
special-case entirely.

The "advice.objectNameWarning" configuration key is rather ugly, and
inconsistent with "advice.objectAmbiguous", but as this configuration
key has always been shown as part of the warning message let's keep it
as-is.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>