Makefile: make ".depend" files *.o-agnostic, use them for *.s
Change the rules added when "COMPUTE_HEADER_DEPENDENCIES" was
introduced in [1] so that we create e.g. a .build/dep/grep.mak file
with the Makefile snippets that declare what "grep.o" depends on,
rather than having a ".depend/grep.o.d" file.

Then in this new file create dependency targets not only for "grep.o",
but also "grep.s". We can use the latter to solve a long-standing
limitation of the "make <c-file-basename>.s" rule.

When it was added in [2] it and the *.o rule depended on a wildcard
of *.h files, but there was no such rule put in place for the new *.s
files, [3] then worked around this limitation by having it depend on
"FORCE". I.e. to always re-generate it.

What we could ideally have done instead is to use the
COMPUTE_HEADER_DEPENDENCIES facility added in [4] to create a
dependency graph for the *.s files, which is what this change does.

I.e. if we've compiled the *.o or *.s file before, have Makefile use
the dependency graph generated by the compiler.

Since we've done that let's add a "asm-obj" convenience target to
create all these *.s files, and fix the long-standing issue of "make
clean" not knowing that it should remove these files.

1. dfea575017d (Makefile: lazily compute header dependencies,
   2010-01-26)
2. e36cb1c16df (Makefile: add an explicit rule for building assembly
   output, 2007-05-30)
3. 31838b4dcd2 (Makefile: regenerate assembler listings when asked,
   2010-01-06)
4. dfea575017d (Makefile: lazily compute header dependencies,
   2010-01-26)

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>