Makefile: add assertion of advice.txt order to "check-docs"
Add a check-advice-docs dependency to the existing "check-docs" rule,
and have that rule assert that the sections in
Documentation/config/advice.txt are in sorted order.

I initially implemented this as:

	diff --git a/generate-advice.sh b/generate-advice.sh
	index 3a92728226d..6adadf90eef 100755
	--- a/generate-advice.sh
	+++ b/generate-advice.sh
	@@ -65,6 +65,13 @@ static struct advice_entry advice_setting[] = {
	 };
	 EOF
	 ;;
	+check-advice-docs)
	+	if test "$(advice_list)" != "$(advice_list | LC_ALL=C sort -u)"
	+	then
	+		echo 'Documentation/config/advice.txt is not sorted!' >&2
	+		exit 1
	+	fi
	+	;;
	 *)
	 	echo "unknown target $1" >&2
	 	exit 1

But I like the pure-Makefile solution better, and by creating a
template in shared.mak it'll be easy to re-use it in the future, and
"make" will execute the individual steps one at a time (and possibly
in parallel):

It's also easier to follow the output and see exactly what fails When
it fails due to the file being unsorted it'll emit:

	$ git clean -dxf; make check-advice-docs
	Removing .build/
	Removing GIT-VERSION-FILE
	GIT_VERSION = 2.35.1.487.gb26d83cdcb9.dirty
	    MKDIR -p .build/check-advice-docs
	    GEN .build/check-advice-docs/actual
	    GEN .build/check-advice-docs/expect
	    CMP .build/check-advice-docs/ok
	--- .build/check-advice-docs/expect     2022-03-15 14:27:32.411569975 +0100
	+++ .build/check-advice-docs/actual     2022-03-15 14:27:32.407570011 +0100
	@@ -31,7 +31,6 @@
	        sequencerRescheduled::
	        sequencerResolveConflict::
	        skippedCherryPicks::
	-       someNewConfig::
	        statusAheadBehind::
	        statusHints::
	        statusUoption::
	@@ -40,3 +39,4 @@
	        suggestDetachingHead::
	        updateSparsePath::
	        waitingForEditor::
	+       someNewConfig::
	make: *** [Makefile:3350: .build/check-advice-docs/ok] Error 1

This piggy-backs on the newly added .build directory and
"mkdir_p_parent_template" template added in 6d17629ec91 (Makefiles:
add and use wildcard "mkdir -p" template, 2021-11-06) to add temporary
files that'll be ignored & cleaned by "make clean".