t/Makefile: use a template for "check-chainlint" rules
Reduce the duplication introduced in the preceding commit by having
the rules for making .build/chainlint/%/X from .build/chainlint/%/Y
defined via templates.

This changes the behavior slightly in that we now pipe everything to
stdin, we could add another parameter to add an optional "<", but this
works too.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>