clone tests: test --recurse alias
Add better tests for how exactly the OPT_ALIAS() for the "recurse"
option added in 5c387428f10 (parse-options: don't emit "ambiguous
option" for aliases, 2019-04-29) behaves.

The tests added to t0040-parse-options.sh in that commit test part of
it (i.e. "--alias"), but not that we support e.g. "--recurs" as a
shorthand for the "--recurse-submodules" option given that the alias
is "--recurse", i.e. that a prefix for the alias can be given that's
not a prefix for the option it's aliased to.

Rather than test that specifically let's test all the cases that work
now, to assert that subsequent changes to the implementation are
intended.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>