parse-options.c: add a parse_options_check_flags()
Add a parse_options_check_flags() for those checks that need to check
the "enum parse_opt_flags", or the "flags" and "opts" in combination.

This changes no behavior, and just moves existing code around, but in
a subsequent commit we'll start checking the per-option flags against
the parse_options() flags, this change makes that subsequent behavior
change smaller.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>