builtin tests: split up the different logical -h tests
As discussed on-list[1] when we encounter a BUG() or optbug() in
parse-options.c the output from t0012-help.sh is pretty useless, since
we redirect all out the output to the "output" file.

Let's instead split the test so that we first check if "-h" returns a
status of 129, and then test the output. Let's also split up the
GIT_CEILING_DIRECTORIES test added in 87ad07d7354 (t0012: verify that
built-ins handle `-h` even without gitdir, 2022-02-08) into its own
test.

This means that we run 2x as many "$builtin -h" commands when
everything's OK here, which makes this a bit slower, but this test is
pretty cheap, and the improved output is worth it.

1. https://lore.kernel.org/git/220225.86zgme7vxo.gmgdl@evledraar.gmail.com/

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>