parse-options: detect bad "help" for OPTION_GROUP
Let's assert in parse_options_check() that "OPT_GROUP" either gets an
"" argument, or string that starts with an upper-case character.

For the latter case let's then check that the string doesn't end in a
full stop or ":", and adjust a case in builtin/branch.c where a
"group" inconsistently used that style.

This could have been added to the existing "switch" for "opts->type",
but we'll want to add another check for the "default" case here in a
subsequent commit. Doing it this way makes that change smaller.

1. https://lore.kernel.org/git/e1c5a3258263d05530f236c247603c2f342dac85.1645766599.git.gitgitgadget@gmail.com/

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>